M3-4751 Change: Add titles to all top menu icons
- Create a new component, TopMenuIcon, that wraps an icon in a div
and Tooltip (the div is necessary to make the tooltip display).
- Use the new component on all the icons in the top menu (including the user
profile/account dropdown) to display text explaining their purpose.

- Clean up CMR files in the TopMenu and UserMenu
