+++ packages/manager/src/features/TopMenu/Community.tsx
@@ -3,6 +3,7 @@ import * as classNames from 'classnames';
 import Community from 'src/assets/icons/community_nav.svg';
 import Link from 'src/components/Link';
 import { useStyles } from './iconStyles';
+import TopMenuIcon from './TopMenuIcon';
 
 export const Help: React.FC<{ className?: string }> = ({ className }) => {
   const classes = useStyles();
@@ -12,7 +13,9 @@ export const Help: React.FC<{ className?: string }> = ({ className }) => {
       className={classNames(className, { [classes.icon]: true })}
       to="https://linode.com/community"
     >
-      <Community />
+      <TopMenuIcon title="Linode Cloud Community">
+        <Community />
+      </TopMenuIcon>
     </Link>
   );
 };
