+++ packages/manager/src/features/TopMenu/Help.tsx
@@ -2,6 +2,7 @@ import * as React from 'react';
 import HelpIcon from 'src/assets/icons/get_help.svg';
 import Link from 'src/components/Link';
 import { useStyles } from './iconStyles';
+import TopMenuIcon from './TopMenuIcon';
 
 export const Help: React.FC<{}> = _ => {
   const classes = useStyles();
@@ -11,7 +12,9 @@ export const Help: React.FC<{}> = _ => {
       className={classes.icon}
       to="/support"
     >
-      <HelpIcon />
+      <TopMenuIcon title={'Help & Support'}>
+        <HelpIcon />
+      </TopMenuIcon>
     </Link>
   );
 };
