+++ packages/manager/src/features/TopMenu/TopMenu.tsx
@@ -3,125 +3,112 @@ import * as React from 'react';
 import AppBar from 'src/components/core/AppBar';
 import Hidden from 'src/components/core/Hidden';
 import IconButton from 'src/components/core/IconButton';
-import {
-  createStyles,
-  Theme,
-  withStyles,
-  WithStyles
-} from 'src/components/core/styles';
+import { makeStyles, Theme } from 'src/components/core/styles';
 import Toolbar from 'src/components/core/Toolbar';
 import Typography from 'src/components/core/Typography';
-import AddNewMenu from './AddNewMenu';
-import SearchBar from './SearchBar';
-import UserEventsMenu from './UserEventsMenu';
+import AddNewMenu from './AddNewMenu/AddNewMenu_CMR';
+import Community from './Community';
+import Help from './Help';
+import NotificationButton from './NotificationButton';
+import SearchBar from './SearchBar/SearchBar_CMR';
+import TopMenuIcon from './TopMenuIcon';
 import UserMenu from './UserMenu';
-import UserNotificationsMenu from './UserNotificationsMenu';
 
-type ClassNames = 'root' | 'flex' | 'appBar' | 'toolbar' | 'navIconHide';
-
-const styles = (theme: Theme) =>
-  createStyles({
-    root: {
-      color: theme.palette.text.primary
-    },
-    flex: {
-      flex: 1
-    },
-    appBar: {
-      color: theme.palette.text.primary,
-      backgroundColor: theme.bg.topMenu,
-      position: 'relative',
-      paddingRight: '0 !important'
-    },
-    toolbar: {
-      padding: `${theme.spacing(1)}px 0`,
-      [theme.breakpoints.up('md')]: {
-        paddingLeft: theme.spacing(1),
-        paddingRight: theme.spacing(1)
-      },
-      [theme.breakpoints.up('lg')]: {
-        paddingLeft: theme.spacing(3),
-        paddingRight: theme.spacing(3)
-      }
-    },
-    navIconHide: {
-      order: 1,
-      '& > span': {
-        justifyContent: 'flex-start'
-      },
-      '& svg': {
-        width: 32,
-        height: 32
-      },
-      [theme.breakpoints.up('lg')]: {
-        marginLeft: -theme.spacing(2)
-      }
+const useStyles = makeStyles((theme: Theme) => ({
+  appBar: {
+    height: 50,
+    color: theme.palette.text.primary,
+    backgroundColor: theme.cmrBGColors.bgSecondaryActions,
+    position: 'relative',
+    paddingRight: '0 !important',
+    display: 'flex',
+    justifyContent: 'center',
+    flexDirection: 'row'
+  },
+  toolbar: {
+    padding: 0,
+    height: `50px !important`,
+    width: '100%'
+  },
+  communityIcon: {
+    [theme.breakpoints.down(370)]: {
+      ...theme.visually.hidden
     }
-  });
+  }
+}));
 
 interface Props {
+  isSideMenuOpen: boolean;
   openSideMenu: () => void;
   desktopMenuToggle: () => void;
   isLoggedInAsCustomer: boolean;
   username: string;
 }
 
-type PropsWithStyles = Props & WithStyles<ClassNames>;
+type PropsWithStyles = Props;
 
-class TopMenu extends React.Component<PropsWithStyles> {
-  render() {
-    const { classes, openSideMenu, desktopMenuToggle } = this.props;
+const TopMenu: React.FC<PropsWithStyles> = props => {
+  const {
+    isSideMenuOpen,
+    openSideMenu,
+    username,
+    isLoggedInAsCustomer,
+    desktopMenuToggle
+  } = props;
 
-    return (
-      <React.Fragment>
-        {this.props.isLoggedInAsCustomer && (
-          <div
-            style={{
-              backgroundColor: 'pink',
-              padding: '1em',
-              textAlign: 'center'
-            }}
-          >
-            <Typography style={{ fontSize: '1.2em', color: 'black' }}>
-              You are logged in as customer:{' '}
-              <strong>{this.props.username}</strong>
-            </Typography>
-          </div>
-        )}
-        <AppBar className={classes.appBar}>
-          <Toolbar className={classes.toolbar}>
-            <Hidden smDown>
-              <IconButton
-                color="inherit"
-                aria-label="open menu"
-                onClick={desktopMenuToggle}
-                className={classes.navIconHide}
-              >
+  const classes = useStyles();
+
+  const navHoverText = isSideMenuOpen ? 'Collapse menu' : 'Expand menu';
+
+  return (
+    <React.Fragment>
+      {isLoggedInAsCustomer && (
+        <div
+          style={{
+            backgroundColor: 'pink',
+            padding: '1em',
+            textAlign: 'center'
+          }}
+        >
+          <Typography style={{ fontSize: '1.2em', color: 'black' }}>
+            You are logged in as customer: <strong>{username}</strong>
+          </Typography>
+        </div>
+      )}
+      <AppBar className={classes.appBar}>
+        <Toolbar className={classes.toolbar} variant="dense">
+          <Hidden smDown>
+            <IconButton
+              color="inherit"
+              aria-label="open menu"
+              onClick={desktopMenuToggle}
+            >
+              <TopMenuIcon title={navHoverText}>
                 <MenuIcon />
-              </IconButton>
-            </Hidden>
-            <Hidden mdUp>
-              <IconButton
-                color="inherit"
-                aria-label="open menu"
-                onClick={openSideMenu}
-                className={classes.navIconHide}
-              >
+              </TopMenuIcon>
+            </IconButton>
+          </Hidden>
+          <Hidden mdUp>
+            <IconButton
+              color="inherit"
+              aria-label="open menu"
+              onClick={openSideMenu}
+            >
+              <TopMenuIcon title={navHoverText}>
                 <MenuIcon />
-              </IconButton>
-            </Hidden>
-            <SearchBar />
-            <AddNewMenu />
-            <UserMenu />
-            <UserNotificationsMenu />
-            <UserEventsMenu />
-          </Toolbar>
-        </AppBar>
-      </React.Fragment>
-    );
-  }
-}
+              </TopMenuIcon>
+            </IconButton>
+          </Hidden>
+          <AddNewMenu />
+          <SearchBar />
+          <Help />
+          <Community className={classes.communityIcon} />
+          <NotificationButton />
+          <UserMenu />
+        </Toolbar>
+      </AppBar>
+    </React.Fragment>
+  );
+};
 
-const styled = withStyles(styles);
-
-export default styled(TopMenu);
+export default React.memo(TopMenu);
